begin adding authorization
Our "accessToken" is an empty signed message from the user. We look for
this signature in the "Authorization" header and recover the address
from it. This address represents our "authenticated" user.

pull the address from this signature when trying to create a user.

setup hooks for user service authorization.

NOTE: there are some flaws in this way of authenticating which we need
to address