use jwt for authentication and signed messages
Migrated to a more secure authentication scheme. Initial login will be
using a signed random message from the ethereum account.

When a user tries to authenticate for the first time, we will fail and
issue a randome message that they need to sign. This messge is valid for
5 mins.

Upon recieving an authentication request with a valid signature, we will
return a jwt valid for 1 day with the users address. If the user does
not exist, we will create a new user.

This jwt token is the only type of authentication the services will
accept.