reworked TransferQueue
changed to EventQueue, and basing off of txHash. Needed to do this b/c
when normalizing a note where the project has been canceled, it will
generate 2 Transfer events, and we were previously only using the queue
if this was a newDonation.

Now we add to the queue if the logIndex > 0, and purge the queue by
txHash, 1 fn at a time from admins.ddGiver and Pleges.transfer methods.
This should cover current use cases for liquidPledging